+++ PIR/include/ModulePrinter.hpp
@@ -42,6 +42,7 @@ namespace pcit::pir{
 			auto print_expr_stmt(Expr expr) -> void;
 
 			auto print_meta_file(const meta::File& file) -> void;
+			auto print_meta_subscope(const meta::Subscope& subscope) -> void;
 			auto print_meta_basic_type(const meta::BasicType& type) -> void;
 			auto print_meta_qualified_type(const meta::QualifiedType& qualified_type) -> void;
 			auto print_meta_struct_type(const meta::StructType& struct_type) -> void;
@@ -53,6 +54,7 @@ namespace pcit::pir{
 			auto print_meta_file_id(meta::File::ID meta_file_id) -> void;
 			auto print_source_location(const std::optional<meta::SourceLocation>& source_location) -> void;
 			auto print_meta_scope(const meta::Scope& scope) -> void;
+			auto print_meta_local_scope(const meta::LocalScope& scope) -> void;
 
 			auto print_function_call_impl(
 				const evo::Variant<FunctionID, ExternalFunctionID, PtrCall>& call_target,